Refactored ProjectSelectionPresenter
